TGP-1971: pr changes
